requirements.txt
fastapi
uvicorn
langchain
langchain-huggingface
python-dotenv
